refactor(test): migrate run creation helpers to db-test-seeders
Move DB-direct run creation helpers from api-test-helpers/runs.ts to
db-test-seeders/runs.ts with @why-db-direct justification for each
function. This separates concerns: api-test-helpers now only contains
API-based and service-based helpers, while db-test-seeders holds
functions that must bypass the API due to PostgreSQL timestamp control,
cross-org insertion, or stale state simulation.

Changes:
- Create db-test-seeders/runs.ts with seedTestRun, seedCompletedTestRun,
  seedStalePendingRun, seedOrphanTestRun, and getOrgIdFromVersion
- Remove createTestRunInDb, createTestRunDirect, createOrphanTestRun,
  createCompletedTestRun, insertStalePendingRun, getOrgIdFromVersion,
  and createTestRunInStatus from api-test-helpers/runs.ts
- Update 52 test files to import from db-test-seeders/runs
- Update api-test-helpers/runner.ts to import getOrgIdFromVersion
  from db-test-seeders/runs

Part of #9339 (parent: #9321)
