+++ turbo/apps/web/app/api/zero/runs/queue/__tests__/route.test.ts
@@ -4,7 +4,6 @@ import {
   createTestRequest,
   createTestOrg,
   createTestCompose,
-  createTestRunInDb,
 } from "../../../../../../src/__tests__/api-test-helpers";
 import {
   testContext,
@@ -12,6 +11,7 @@ import {
 } from "../../../../../../src/__tests__/test-helpers";
 import { mockClerk } from "../../../../../../src/__tests__/clerk-mock";
 import { generateSandboxToken } from "../../../../../../src/lib/auth/sandbox-token";
+import { seedTestRun } from "../../../../../../src/__tests__/db-test-seeders/runs";
 
 const context = testContext();
 
@@ -53,7 +53,7 @@ describe("GET /api/zero/runs/queue", () => {
     const userId = uniqueId("zqueue-run");
     await setupOrg(userId);
     const compose = await createTestCompose(`agent-${uniqueId("zqueue")}`);
-    await createTestRunInDb(userId, compose.composeId, {
+    await seedTestRun(userId, compose.composeId, {
       status: "running",
     });
