+++ turbo/apps/web/src/lib/zero/org/__tests__/org-deletion-service.test.ts
@@ -2,7 +2,6 @@ import { describe, it, expect } from "vitest";
 import { testContext, uniqueId } from "../../../../__tests__/test-helpers";
 import {
   createTestCompose,
-  createTestRunInDb,
   createTestSecret,
   createTestVariable,
   createTestAgentSession,
@@ -30,6 +29,7 @@ import {
 } from "../../../../__tests__/api-test-helpers";
 import { createTestZeroAgent } from "../../../../__tests__/db-test-seeders/agents";
 import { deleteOrgData } from "../org-deletion-service";
+import { seedTestRun } from "../../../../__tests__/db-test-seeders/runs";
 
 const context = testContext();
 
@@ -48,17 +48,17 @@ describe("deleteOrgData", () => {
     const { userId, orgId } = await context.setupUser();
 
     const { composeId } = await createTestCompose("cancel-test");
-    const { runId: queuedRunId } = await createTestRunInDb(userId, composeId, {
+    const { runId: queuedRunId } = await seedTestRun(userId, composeId, {
       status: "queued",
     });
-    const { runId: pendingRunId } = await createTestRunInDb(userId, composeId, {
+    const { runId: pendingRunId } = await seedTestRun(userId, composeId, {
       status: "pending",
     });
-    const { runId: runningRunId } = await createTestRunInDb(userId, composeId, {
+    const { runId: runningRunId } = await seedTestRun(userId, composeId, {
       status: "running",
       startedAt: new Date(),
     });
-    await createTestRunInDb(userId, composeId, {
+    await seedTestRun(userId, composeId, {
       status: "completed",
       completedAt: new Date(),
     });
@@ -91,7 +91,7 @@ describe("deleteOrgData", () => {
     const { userId, orgId } = await context.setupUser();
 
     const { composeId } = await createTestCompose("cascade-run-test");
-    const { runId } = await createTestRunInDb(userId, composeId, {
+    const { runId } = await seedTestRun(userId, composeId, {
       status: "completed",
       completedAt: new Date(),
     });
@@ -210,7 +210,7 @@ describe("deleteOrgData", () => {
     // Composes, sessions, runs
     const { composeId, agentId } = await createTestCompose("full-org-test");
     const session = await createTestAgentSession(userId, composeId);
-    const { runId } = await createTestRunInDb(userId, composeId, {
+    const { runId } = await seedTestRun(userId, composeId, {
       status: "running",
       startedAt: new Date(),
     });
@@ -327,7 +327,7 @@ describe("deleteOrgData", () => {
     const { userId, orgId } = await context.setupUser();
 
     const { composeId } = await createTestCompose("idempotent-test");
-    await createTestRunInDb(userId, composeId, {
+    await seedTestRun(userId, composeId, {
       status: "running",
       startedAt: new Date(),
     });
